Skip build directory in DeleteDirectory.

+++ source/Nuke.Core/IO/FileSystemTasks.cs
@@ -65,6 +65,12 @@ namespace Nuke.Core.IO
             if (!Directory.Exists(directory))
                 return;
 
+            if (PathConstruction.IsDescendantPath(EnvironmentInfo.BuildDirectory, directory))
+            {
+                Logger.Warn($"Skipping directory '{directory}' because it is contained in the build project directory...");
+                return;
+            }
+
             Logger.Info($"Deleting directory '{directory}'...");
             DeleteDirectoryInternal(directory);
         }
